intellidoc-ai-service/requirements.txt
fastapi # //Use for building API 
uvicorn # //Use for running API 
google-genai # //Use for AI model (new unified SDK) 
python-dotenv # //Use for environment variables 
pydantic # //Use for type checking and data validation 
